LPS-127934 extract getDDMSettingsContextWithVisualProperties to dataConverter
- From DataLayoutBuilder to dataConverter
