+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/utils/normalizers.es.js
@@ -12,6 +12,8 @@
  * details.
  */
 
+import {DataConverter} from 'data-engine-taglib';
+
 import {getValidName} from '../utils/utils.es';
 
 export const normalizeNames = ({
@@ -65,9 +67,15 @@ export const normalizeDataLayout = ({
 		}
 	});
 
-	dataDefinitionFields.forEach((definitionField) => {
-		const fieldProperties = dataLayoutBuilder.getDDMSettingsContextWithVisualProperties(
-			definitionField
+	dataDefinitionFields.forEach((dataDefinitionField) => {
+		const {editingLanguageId} = dataLayoutBuilder.getState();
+		const fieldTypes = dataLayoutBuilder.getFieldTypes();
+		const fieldProperties = DataConverter.getDDMSettingsContextWithVisualProperties(
+			{
+				dataDefinitionField,
+				editingLanguageId,
+				fieldTypes,
+			}
 		);
 
 		// Ignore this visual properties because it is treated differently
@@ -75,8 +83,8 @@ export const normalizeDataLayout = ({
 		delete fieldProperties['label'];
 		delete fieldProperties['required'];
 
-		dataLayoutFields[definitionField.name] = {
-			...dataLayoutFields[definitionField.name],
+		dataLayoutFields[dataDefinitionField.name] = {
+			...dataLayoutFields[dataDefinitionField.name],
 			...fieldProperties,
 		};
 	});
