+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/utils/dataConverter.es.js
@@ -262,6 +262,33 @@ export function getDataDefinitionFieldByFieldName({
 	};
 }
 
+export function getDDMSettingsContextWithVisualProperties({
+	dataDefinitionField,
+	editingLanguageId,
+	fieldTypes,
+}) {
+	const {pages} = getDDMFormFieldSettingsContext({
+		dataDefinitionField,
+		editingLanguageId,
+		fieldTypes,
+	});
+	const visitor = new PagesVisitor(pages);
+
+	const fieldProperties = {};
+
+	visitor.mapFields(
+		({fieldName, localizable, localizedValue, value, visualProperty}) => {
+			if (visualProperty) {
+				fieldProperties[fieldName] = localizable
+					? localizedValue
+					: value;
+			}
+		}
+	);
+
+	return fieldProperties;
+}
+
 /**
  * Converts a FieldSet from data-engine to form-builder data definition
  */
